refactor folder structure for app/
