+++ frontend/src/modules/alarms/modal/DiscardAlarmLimitsContent.tsx
@@ -6,13 +6,13 @@
 import { Grid, makeStyles, Theme, Typography } from '@material-ui/core';
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { alarmConfiguration, AlarmConfiguration } from '../alarms/limits/AlarmLimitsPage';
-import { VentilationMode, Range } from '../../store/proto/mcu_pb';
+import { alarmConfiguration, AlarmConfiguration } from '../limits/AlarmLimitsPage';
+import { VentilationMode, Range } from '../../../store/proto/mcu_pb';
 import {
   getAlarmLimitsRequestDraft,
   getAlarmLimitsRequest,
   getAlarmLimitsUnsavedKeys,
-} from '../../store/controller/selectors';
+} from '../../../store/controller/selectors';
 
 const useStyles = makeStyles((theme: Theme) => ({
   marginContent: {
