+++ frontend/src/modules/alarms/modal/index.ts
@@ -1,0 +1,2 @@
+export * from './DiscardAlarmLimitsContent';
+export { AlarmModal, AlarmModalContent } from './AlarmModal';
