+++ frontend/src/modules/app/landing-page/LandingPage.tsx
@@ -5,7 +5,7 @@
 import { Grid } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 import React from 'react';
-import Carousel from '../displays/Carousel';
+import Carousel from '../../displays/Carousel';
 
 const useStyles = makeStyles(() => ({
   slideHead: {
