+++ frontend/src/modules/dashboard/components/ValueInfo.tsx
@@ -12,7 +12,7 @@ import React, { useEffect, useState } from 'react';
 import { shallowEqual, useSelector } from 'react-redux';
 import { getAlarmLimitsRequest } from '../../../store/controller/selectors';
 import { setMultiPopupOpen } from '../../app/Service';
-import { AlarmModal } from '../../controllers';
+import { AlarmModal } from '../../alarms/modal/AlarmModal';
 import { Range } from '../../../store/proto/mcu_pb';
 import { SelectorType, ValueSelectorDisplay } from '../../displays/ValueSelectorDisplay';
