+++ frontend/src/modules/navigation/Routes.tsx
@@ -21,11 +21,11 @@ import ScreensaverRoute from './layouts/ScreensaverRoute';
 import SidebarRoute from './layouts/SidebarRoute';
 import SidebarSlideRoute from './layouts/SidebarSlideRoute';
 import DashboardPage from '../dashboard/DashboardPage';
-import LandingPage from '../landing-page/LandingPage';
+import LandingPage from '../app/landing-page/LandingPage';
 import LogsPage from '../logs/LogsPage';
 import ModesPage from '../modes/ModesPage';
-import QuickStartPage from '../quickstart/QuickStartPage';
-import ScreensaverPage from '../screensaver/ScreensaverPage';
+import QuickStartPage from '../app/quickstart/QuickStartPage';
+import ScreensaverPage from '../app/screensaver/ScreensaverPage';
 import SettingsPage from '../settings/SettingsPage';
 import {
   ALARMS_ROUTE,
