--- frontend/src/modules/overlay/index.ts
@@ -1,1 +1,0 @@
-export * from './OverlayScreen';
