+++ frontend/src/modules/toolbar/StartPauseButton.tsx
@@ -23,9 +23,10 @@ import {
   Range,
 } from '../../store/proto/mcu_pb';
 import { MessageType } from '../../store/proto/types';
-import { ModalPopup, DiscardAlarmLimitsContent } from '../modals';
+import { ModalPopup } from '../modals';
 import { QUICKSTART_ROUTE, DASHBOARD_ROUTE } from '../navigation/constants';
 import { getAlarmLimitsModalPopup, setAlarmLimitsModalPopup } from '../app/Service';
+import { DiscardAlarmLimitsContent } from '../alarms/modal';
 
 const StartPauseButton = ({ staticStart }: { staticStart?: boolean }): JSX.Element => {
   const dispatch = useDispatch();
